requirements.txt
djangorestframework
django
pydantic
pytest
